Moved rustpkg::CtxMethods::run(cmd: Command ...) to rustpkg::run_cmd.rs
